+++ lib/datamappify/repository/query_method/callbacks.rb
@@ -8,7 +8,8 @@ module Datamappify
           klass.class_eval do
             include Hooks
 
-            define_hooks :before_load,    :after_load,
+            define_hooks :before_init,    :after_init,
+                         :before_load,    :after_load,
                          :before_find,    :after_find,
                          :before_create,  :after_create,
                          :before_update,  :after_update,