+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Proper dirty tracking in record updating, fixes an issue of setting an attribute to `nil`.
 - Added `has_one` association.
 - Added `before_load`, `after_load`, `before_find` and `after_find` callbacks.
+- Added `before_init` and `after_init` callbacks.
 
 ## 0.70.0.beta1 [2013-09-26]